Move logging setup to common ground
