+++ crmprtd/ec/__init__.py
@@ -22,23 +22,6 @@ ns = {
 }
 
 
-def logging_setup(log_conf, log, error_email, log_level):
-    with open(log_conf, 'rb') as f:
-        log_conf = yaml.load(f)
-    if log:
-        log_conf['handlers']['file']['filename'] = log
-    else:
-        log = log_conf['handlers']['file']['filename']
-    if error_email:
-        log_conf['handlers']['mail']['toaddrs'] = error_email
-    logging.config.dictConfig(log_conf)
-    log = logging.getLogger('crmprtd.ec')
-    if log_level:
-        log.setLevel(log_level)
-
-    return log
-
-
 def makeurl(freq='daily', province='BC', language='e', time=datetime.utcnow()):
     """Construct a URL for fetching the data file
     freq: daily|hourly
