+++ crmprtd/moti/__init__.py
@@ -13,22 +13,6 @@ from crmprtd import Timer
 log = logging.getLogger(__name__)
 
 
-def logging_setup(log_conf, log, error_email, log_level):
-    log_c = yaml.load(log_conf)
-    if log:
-        log_c['handlers']['file']['filename'] = log
-    else:
-        log = log_c['handlers']['file']['filename']
-    if error_email:
-        log_c['handlers']['mail']['toaddrs'] = error_email
-    logging.config.dictConfig(log_c)
-    log = logging.getLogger('crmprtd.moti')
-    if log_level:
-        log.setLevel(log_level)
-
-    return log
-
-
 xsl = resource_filename('crmprtd', 'data/moti.xsl')
 transform = XSLT(parse(xsl))
