+++ crmprtd/wamr/__init__.py
@@ -150,25 +150,6 @@ class DataLogger(object):
             yield row
 
 
-def setup_logging(level, filename=None, email=None):
-    '''Read in the logging configuration and return a logger object
-    '''
-    log_conf = yaml.load(resource_stream('crmprtd', '/data/logging.yaml'))
-    if filename:
-        log_conf['handlers']['file']['filename'] = filename
-    else:
-        filename = log_conf['handlers']['file']['filename']
-    if email:
-        log_conf['handlers']['mail']['toaddrs'] = email
-    logging.config.dictConfig(log_conf)
-    log = logging.getLogger('crmprtd.wamr')
-
-    if level:
-        log.setLevel(level)
-
-    return log
-
-
 def rows2db(sesh, rows, error_file, log, diagnostic=False):
     '''
     Args:
